pyproject.toml
[project]
name = "ideas-pipeline"
version = "0.1.0"
description = "CC idea-capture and promote pipeline — writes to ~/Documents/ObsidianVault/"
requires-python = ">=3.12"
dependencies = [
  "click>=8.1",
  "pyyaml>=6.0",
  "python-dateutil>=2.8",
]

[project.scripts]
ideas = "ideas.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["ideas*"]
